requirements.txt
asgiref==3.4.1
autopep8==1.6.0
cx-Oracle==8.3.0
Django==4.0
django-axes==5.31.0
django-brutebuster==0.1.8
django-brutebuster2==0.2.0
django-crispy-forms==1.13.0
django-ipware==4.0.2
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-smart-selects @ git+https://github.com/digi604/django-smart-selects.git@1ceafd8ff63e5888d5c0bcb9ff6f5d4e03769fcf
django-widget-tweaks==1.4.12
fontawesomefree==5.15.4
jsonfield==3.1.0
Pillow==9.0.0
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
swapper==1.3.0
toml==0.10.2
tzdata==2021.5
